requirements.txt
paddlepaddle-gpu>=2.0.0
visualdl>=2.1.1